split up test_suite.py somewhat
