+++ corehq/apps/app_manager/tests/test_suite.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from corehq.apps.app_manager.exceptions import CaseXPathValidationError
 import re
 from django.test import SimpleTestCase
 from corehq.apps.app_manager.const import APP_V2
@@ -30,9 +29,6 @@ from corehq.apps.app_manager.models import (
 from corehq.apps.app_manager.tests.app_factory import AppFactory
 from corehq.apps.app_manager.tests.util import TestXmlMixin, commtrack_enabled
 from corehq.apps.app_manager.xpath import (
-    dot_interpolate,
-    UserCaseXPath,
-    interpolate_xpath,
     session_var,
 )
 from corehq.toggles import NAMESPACE_DOMAIN
@@ -41,7 +37,6 @@ from toggle.shortcuts import update_toggle_cache, clear_toggle_cache
 
 from lxml import etree
 import commcare_translations
-from mock import patch
 
 
 class SuiteTest(SimpleTestCase, TestXmlMixin):
@@ -684,112 +679,3 @@ class SuiteTest(SimpleTestCase, TestXmlMixin):
             'reports.ip1bjs8xtaejnhfrbzj2r6v1fi6hia4i=CommBugz',
             app.create_app_strings('default'),
         )
-
-
-class RegexTest(SimpleTestCase):
-
-    def test_regex(self):
-        replacement = "@case_id stuff"
-        cases = [
-            ('./lmp < 570.5', '%s/lmp < 570.5'),
-            ('stuff ./lmp < 570.', 'stuff %s/lmp < 570.'),
-            ('.53 < hello.', '.53 < hello%s'),
-        ]
-        for case in cases:
-            self.assertEqual(
-                dot_interpolate(case[0], replacement),
-                case[1] % replacement
-            )
-
-    def test_interpolate_xpath(self):
-        replacements = {
-            'case': "<casedb stuff>",
-            'user': UserCaseXPath().case(),
-            'session': "instance('commcaresession')/session",
-        }
-        cases = [
-            ('./lmp < 570.5', '{case}/lmp < 570.5'),
-            ('#case/lmp < 570.5', '{case}/lmp < 570.5'),
-            ('stuff ./lmp < 570.', 'stuff {case}/lmp < 570.'),
-            ('stuff #case/lmp < 570.', 'stuff {case}/lmp < 570.'),
-            ('.53 < hello.', '.53 < hello{case}'),
-            ('.53 < hello#case', '.53 < hello{case}'),
-            ('#session/data/username', '{session}/data/username'),
-            ('"jack" = #session/username', '"jack" = {session}/username'),
-            ('./@case_id = #session/userid', '{case}/@case_id = {session}/userid'),
-            ('#case/@case_id = #user/@case_id', '{case}/@case_id = {user}/@case_id'),
-            ('#host/foo = 42', "instance('casedb')/casedb/case[@case_id={case}/index/host]/foo = 42"),
-            ("'ham' = #parent/spam", "'ham' = instance('casedb')/casedb/case[@case_id={case}/index/parent]/spam"),
-        ]
-        for case in cases:
-            self.assertEqual(
-                interpolate_xpath(case[0], replacements['case']),
-                case[1].format(**replacements)
-            )
-
-    def test_interpolate_xpath_error(self):
-        for case in ('./lmp < 570.5', '#case/lmp < 570.5'):
-            with self.assertRaises(CaseXPathValidationError):
-                interpolate_xpath(case, None),
-
-
-class FormFilterErrorTests(SimpleTestCase, TestXmlMixin):
-    file_path = ('data', 'suite')
-
-    def setUp(self):
-        self.suite_xml_is_usercase_in_use_patch = patch(
-            'corehq.apps.app_manager.suite_xml.sections.menus.is_usercase_in_use'
-        )
-        self.suite_xml_is_usercase_in_use_mock = self.suite_xml_is_usercase_in_use_patch.start()
-        self.factory = AppFactory(build_version='2.9')
-
-    def tearDown(self):
-        self.suite_xml_is_usercase_in_use_patch.stop()
-
-    def test_error_when_no_case(self):
-        self.suite_xml_is_usercase_in_use_mock.return_value = True
-
-        __, reg_form = self.factory.new_basic_module('reg_module', 'mother')
-        self.factory.form_opens_case(reg_form)
-        reg_form.form_filter = './due_date <= today()'
-
-        with self.assertRaises(CaseXPathValidationError):
-            self.factory.app.create_suite()
-
-    def test_no_error_when_user_case(self):
-        self.suite_xml_is_usercase_in_use_mock.return_value = True
-
-        __, reg_form = self.factory.new_basic_module('reg_module', 'mother')
-        self.factory.form_opens_case(reg_form)
-        reg_form.form_filter = '#user/due_date <= today()'
-
-        expected = """
-        <partial>
-            <menu id="m0">
-            <text>
-              <locale id="modules.m0"/>
-            </text>
-            <command id="m0-f0" relevant="instance('casedb')/casedb/case[@case_type='commcare-user'][hq_user_id=instance('commcaresession')/session/context/userid][1]/due_date &lt;= today()"/>
-          </menu>
-        </partial>
-        """
-        self.assertXmlPartialEqual(expected, self.factory.app.create_suite(), './menu')
-
-    def test_no_error_when_case(self):
-        self.suite_xml_is_usercase_in_use_mock.return_value = False
-
-        __, update_form = self.factory.new_basic_module('update_mother', 'mother')
-        self.factory.form_requires_case(update_form)
-        update_form.form_filter = '#case/due_date <= today()'
-
-        expected = """
-        <partial>
-          <menu id="m0">
-            <text>
-              <locale id="modules.m0"/>
-            </text>
-            <command id="m0-f0" relevant="instance('casedb')/casedb/case[@case_id=instance('commcaresession')/session/data/case_id]/due_date &lt;= today()"/>
-          </menu>
-        </partial>
-        """
-        self.assertXmlPartialEqual(expected, self.factory.app.create_suite(), './menu')
